chromium/third_party/blink/web_tests/virtual/out-of-blink-cspee/external/wpt/content-security-policy/embedded-enforcement/subsumption_algorithm-nonces-expected.txt

This is a testharness.js-based test.
PASS Exact nonce subsumes.
PASS Any nonce subsumes.
FAIL A nonce has to be returned if required by the embedder. assert_equals: expected (undefined) undefined but got (boolean) true
PASS Multiples nonces returned subsume.
PASS Nonce intersection is still done on exact match - non-matching nonces.
FAIL Nonce intersection is still done on exact match - matching nonces. assert_equals: expected (undefined) undefined but got (boolean) true
PASS Other expressions still have to be subsumed - positive test.
FAIL Other expressions still have to be subsumed - negative test assert_unreached: No message should be sent from the frame. Reached unreachable code
Harness: the test ran to completion.
